+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v3.2.5] - 2026-03-19
+
+### Changed
+
+- Updated [@adguard/agtree] to v4.0.2.
+- Updated [@adguard/scriptlets] to v2.3.0.
+- Updated [@adguard/tsurlfilter] to v4.0.3.
+
+[v3.2.5]: https://github.com/AdguardTeam/FiltersCompiler/compare/v3.2.4...v3.2.5
+
 ## [v3.2.4] - 2025-12-16
 
 ### Changed